requirements.txt
nose
sklearn-crfsuite
neo4j-driver
